requirements.txt
aiohttp>=3.9.3
aiosignal>=1.3.1
anyio>=4.3.0
attrs>=23.2.0
Brotli>=1.1.0
certifi>=2024.2.2
cffi>=1.16.0
charset-normalizer>=3.3.2
frozenlist>=1.4.1
h11>=0.14.0
httpcore>=1.0.5
httpx>=0.27.0
idna>=3.6
multidict>=6.0.5
mutagen>=1.47.0
py-cord>=2.5.0
pycparser>=2.22
pycryptodomex>=3.20.0
PyNaCl>=1.5.0
python-dotenv>=1.0.1
requests>=2.31.0
sniffio>=1.3.1
urllib3>=2.2.1
websockets>=12.0
yarl>=1.9.4
youtube-dl>=2021.12.17
youtube-search-python>=1.6.6
yt-dlp>=2024.3.10
